% Auto-generated: do not edit by hand
\name{dashLottieLoading}

\alias{dashLottieLoading}

\title{DashLottieLoading component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
dashLottieLoading(id=NULL, height=NULL, path=NULL, width=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{height}{Character. The height of the lottie when this component is rendered.}

\item{path}{Character. The path of the lottie to show when this component is rendered.}

\item{width}{Character. The width of the lottie when this component is rendered.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

